+++ nanoworks/locales/de.py
@@ -7,7 +7,10 @@ STRINGS = {
     "fig_band_ylabel": "Energie (eV)",
     
     # Optical Properties Figure
-    "fig_optical_energy_x": "Energie (eV)",
+    "fig_optical_xlabel": "Energie (eV)",
+    "fig_optical_dielectric": "Dielektrische Funktion",
+    "fig_optical_abs": r"Absorptionskoeffizient ($cm^{-1}$)",
+    "fig_optical_ref": "Reflektivität",
     
     # Elastic/Phonon Figures
     "fig_phonon_ylabel": "Frequenz (THz)"