+++ nanoworks/locales/ru.py
@@ -7,7 +7,10 @@ STRINGS = {
     "fig_band_ylabel": "Энергия (эВ)",
     
     # Optical Properties Figure
-    "fig_optical_energy_x": "Энергия (эВ)",
+    "fig_optical_xlabel": "Энергия (эВ)",
+    "fig_optical_dielectric": "Диэлектрическая функция",
+    "fig_optical_abs": r"Коэффициент поглощения ($cm^{-1}$)",
+    "fig_optical_ref": "Коэффициент отражения",
     
     # Elastic/Phonon Figures
     "fig_phonon_ylabel": "Частота (ТГц)"